Add more combination for query customizer
